1.0.1 - Fix issue when using relative directories
The 1.0.0 version will only correctly recurse through parent directories
if you pass a full path in as the second parameter:

    check_777 /full/path/to/directory

If you pass a relative path (`check_777 .`) it enters an infinite loop.

The functions within the script now convert paths to absolute paths
using `readlink -f` so that users of the program can use any valid path
to use it.